* feature/dev_simple_recyclerview:
  chore: DELETE 방식 좋아요 API 로직 수정
  feat: Controller 와 ViewHolder 간의 읜존성을 없애기 위한 방법 - ViewHolder 에서 EntryPoint 지정하여 GoodsRepository 추가 - data Module 에 RxBusEvent 추가
  feat: Controller 와 ViewHolder 간의 읜존성을 없애기 위한 방법 고민.. - ViewHolder 에서 Inject 할수 있는 방안을 고민해보자!

+++ android/model/src/main/java/com/til/model/goods/GoodsEntity.kt
@@ -1,8 +1,6 @@
 package com.til.model.goods
 
-import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.Serializer
 
 /**
  * Description :
@@ -11,7 +9,7 @@ import kotlinx.serialization.Serializer
  */
 @Serializable
 data class GoodsEntity(
-    val id: Int = 0,
+    val id: Long = 0,
     val title: String = "",
     val message: String = "",
     val imagePath: String = ""
